refactor(ui): split api.ts, secure credential storage
